Use more abstract factories for documents
